+++ packages/af-utils-react-virtual-headless/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/react-virtual-headless
 
+## 0.0.28
+
+### Patch Changes
+
+-   optimized getIndex
+
 ## 0.0.27
 
 ### Patch Changes